requirements.txt
jinja2
yasha
python-dateutil
click
psycopg2 
